+++ sdk/elastic/arm-elastic/src/api/vMHost/index.ts
@@ -2,4 +2,4 @@
 // Licensed under the MIT License.
 
 export { list } from "./operations.js";
-export type { VMHostListOptionalParams } from "./options.js";
+export type { vMHostListOptionalParams } from "./options.js";