+++ sdk/elastic/arm-elastic/src/api/vMIngestion/index.ts
@@ -2,4 +2,4 @@
 // Licensed under the MIT License.
 
 export { details } from "./operations.js";
-export type { VMIngestionDetailsOptionalParams } from "./options.js";
+export type { vMIngestionDetailsOptionalParams } from "./options.js";