+++ sdk/elastic/arm-elastic/src/classic/vMIngestion/index.ts
@@ -2,32 +2,32 @@
 // Licensed under the MIT License.
 
 import type { MicrosoftElasticContext } from "../../api/microsoftElasticContext.js";
-import { details } from "../../api/vmIngestion/operations.js";
-import type { VMIngestionDetailsOptionalParams } from "../../api/vmIngestion/options.js";
+import { details } from "../../api/vMIngestion/operations.js";
+import type { vMIngestionDetailsOptionalParams } from "../../api/vMIngestion/options.js";
 import type { VMIngestionDetailsResponse } from "../../models/models.js";
 
-/** Interface representing a VMIngestion operations. */
-export interface VMIngestionOperations {
+/** Interface representing a vMIngestion operations. */
+export interface vMIngestionOperations {
   /** List detailed information about VM ingestion that will be monitored by the Elastic monitor resource, ensuring optimal observability and performance. */
   details: (
     resourceGroupName: string,
     monitorName: string,
-    options?: VMIngestionDetailsOptionalParams,
+    options?: vMIngestionDetailsOptionalParams,
   ) => Promise<VMIngestionDetailsResponse>;
 }
 
-function _getVMIngestion(context: MicrosoftElasticContext) {
+function _getvMIngestion(context: MicrosoftElasticContext) {
   return {
     details: (
       resourceGroupName: string,
       monitorName: string,
-      options?: VMIngestionDetailsOptionalParams,
+      options?: vMIngestionDetailsOptionalParams,
     ) => details(context, resourceGroupName, monitorName, options),
   };
 }
 
-export function _getVMIngestionOperations(context: MicrosoftElasticContext): VMIngestionOperations {
+export function _getvMIngestionOperations(context: MicrosoftElasticContext): vMIngestionOperations {
   return {
-    ..._getVMIngestion(context),
+    ..._getvMIngestion(context),
   };
 }